refactor: 책 추천 컴포넌트 경로 변경 적용
