+++ src/components/BookRecommend/BookRecommendCard.tsx
@@ -1,5 +1,5 @@
 import type { RecommendationDto } from "../../types/dto";
-import BookRecommendHeader from "./BookRecommendHeader";
+import BookRecommendHeader from "../../pages/BookRecommend/BookRecommendHeader";
 import StarRating from "./StarRating";
 
 interface BookRecommendCardProps {
